requirements.txt
python-dotenv==1.1.0
solana==0.36.6
solders==0.26.0
requests==2.32.3
base58==2.1.1
psycopg2-binary==2.9.10
